+++ rotkehlchen/api/rest.py
@@ -1770,6 +1770,25 @@ class RestAPI():
         )
 
     @require_premium_user(active_check=False)
+    def get_uniswap_events_history(
+            self,
+            async_query: bool,
+            reset_db_data: bool,
+            from_timestamp: Timestamp,
+            to_timestamp: Timestamp,
+    ) -> Response:
+        return self._api_query_for_eth_module(
+            async_query=async_query,
+            module='uniswap',
+            method='get_events_history',
+            query_specific_balances_before=None,
+            addresses=self.rotkehlchen.chain_manager.queried_addresses_for_module('uniswap'),
+            reset_db_data=reset_db_data,
+            from_timestamp=from_timestamp,
+            to_timestamp=to_timestamp,
+        )
+
+    @require_premium_user(active_check=False)
     def get_uniswap_trades_history(
             self,
             async_query: bool,