+++ rotkehlchen/api/v1/resources.py
@@ -993,6 +993,26 @@ class UniswapBalancesResource(BaseResource):
         return self.rest_api.get_uniswap_balances(async_query=async_query)
 
 
+class UniswapEventsHistoryResource(BaseResource):
+
+    get_schema = AsyncHistoricalQuerySchema()
+
+    @use_kwargs(get_schema, location='json_and_query')  # type: ignore
+    def get(
+            self,
+            async_query: bool,
+            reset_db_data: bool,
+            from_timestamp: Timestamp,
+            to_timestamp: Timestamp,
+    ) -> Response:
+        return self.rest_api.get_uniswap_events_history(
+            async_query=async_query,
+            reset_db_data=reset_db_data,
+            from_timestamp=from_timestamp,
+            to_timestamp=to_timestamp,
+        )
+
+
 class UniswapTradesHistoryResource(BaseResource):
 
     get_schema = AsyncHistoricalQuerySchema()