+++ rotkehlchen/chain/ethereum/uniswap/graph.py
@@ -102,3 +102,91 @@ SWAPS_QUERY = (
     }}}}
     """
 )
+
+# Get LP events queries
+# TODO: At the moment there are no protocol fees. However, it is possible they
+# turn them on in a future. The fees (from `feeTo` field in both `mints` and
+# `burns` schemas) would be already factorized in the events amounts.
+# Requesting and storing them in DB would just be for informing the user.
+# https://uniswap.org/docs/v2/advanced-topics/fees/#protocol-fees
+MINTS_QUERY = (
+    """
+    mints
+    (
+        first: $limit,
+        skip: $offset,
+        where: {{
+            to: $address,
+            timestamp_gte: $start_ts,
+            timestamp_lte: $end_ts,
+        }}
+    ) {{
+        transaction {{
+            id
+        }}
+        logIndex
+        timestamp
+        to
+        pair {{
+            id
+            token0 {{
+                id
+                decimals
+                name
+                symbol
+            }}
+            token1 {{
+                id
+                decimals
+                name
+                symbol
+            }}
+        }}
+        amount0
+        amount1
+        amountUSD
+        liquidity
+    }}}}
+    """
+)
+
+BURNS_QUERY = (
+    """
+    burns
+    (
+        first: $limit,
+        skip: $offset,
+        where: {{
+            sender: $address,
+            timestamp_gte: $start_ts,
+            timestamp_lte: $end_ts,
+        }}
+    ) {{
+        transaction {{
+            id
+        }}
+        logIndex
+        timestamp
+        sender
+        pair {{
+            id
+            token0 {{
+                id
+                decimals
+                name
+                symbol
+            }}
+            token1 {{
+                id
+                decimals
+                name
+                symbol
+            }}
+        }}
+        amount0
+        amount1
+        amountUSD
+        liquidity
+    }}}}
+    """
+)